requirements.txt
flask
flask_socketio
eventlet
google-cloud-secret-manager
google-cloud-firestore
bcrypt
debugpy